Get rid of ServerRootDescription
