+++ server/src/main/java/org/jboss/as/server/operations/ServerShutdownHandler.java
@@ -32,7 +32,7 @@ import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.OperationStepHandler;
 import org.jboss.as.controller.descriptions.DescriptionProvider;
 import org.jboss.as.process.ExitCodes;
-import org.jboss.as.server.controller.descriptions.ServerRootDescription;
+import org.jboss.as.server.controller.descriptions.ServerDescriptions;
 import org.jboss.dmr.ModelNode;
 
 /**
@@ -75,6 +75,6 @@ public class ServerShutdownHandler implements OperationStepHandler, DescriptionP
 
     /** {@inheritDoc} */
     public ModelNode getModelDescription(final Locale locale) {
-        return ServerRootDescription.getShutdownOperationDescription(locale);
+        return ServerDescriptions.getShutdownOperationDescription(locale);
     }
 }
